Added BookingIDBackend and BookingIDAuthenticationForm
The idea is to re-use Django's AUTHENTICATION_BACKENDS. We would add
our new BookingIDBackend to the list of auth backends. Now we can call Django's
login method which requires a request and a user. It will search for username
and password in the request POST data. Therefore the new
BookingIDAuthenticationForm also has username and password fields, although
it's labels will say "Email" and "Booking ID".